+++ flake.nix
@@ -2,8 +2,7 @@
 # License, v. 2.0. If a copy of the MPL was not distributed with this
 # file, You can obtain one at https://mozilla.org/MPL/2.0/.
 {
-  # TODO: Add description
-  description = "";
+  description = "A Discord bot for listening music from a Subsonic server.";
 
   inputs = {
     nixpkgs.url = "github:nixos/nixpkgs/nixos-unstable";