+++ src/disopy/__init__.py
@@ -16,7 +16,7 @@ APP_NAME: Final[str] = "Disopy"
 APP_NAME_UPPER: Final[str] = APP_NAME.upper()
 APP_NAME_LOWER: Final[str] = APP_NAME.lower()
 
-DEFAULT_CONFIG_PATH: Final[Path] = Path(platformdirs.user_config_dir(APP_NAME))
-DEFAULT_CACHE_PATH: Final[Path] = Path(platformdirs.user_cache_dir(APP_NAME))
+DEFAULT_CONFIG_PATH: Final[Path] = Path(platformdirs.user_config_dir(APP_NAME_LOWER))
+DEFAULT_CACHE_PATH: Final[Path] = Path(platformdirs.user_cache_dir(APP_NAME_LOWER))
 
 colorama.just_fix_windows_console()