+++ CHANGELOG.md
@@ -8,8 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-...
+- The `volume` entry in the config file to set the base volume of all the songs.
 
 ### Changed
+- Bump up the Python version to `3.13.0`.
 - The project is now licensed under the [MPL 2.0](https://www.mozilla.org/en-US/MPL/) license instead of the [MIT](https://opensource.org/license/MIT) license.
 - The project now uses the [Hatch](https://hatch.pypa.io/latest/) Python project manager.